Sort out udesired use of doubles in temp and pressure calcs.

+++ drivers/sensor/abp2s/abp2s_utils.c
@@ -26,14 +26,14 @@ bool abp2s_check_status(uint8_t status)
 /* MCX FPU is single precision */
 float abp2s_calculate_pressure(uint32_t counts)
 {
-    float pressure = ((((float)counts - OUT_MIN_10) * (PMAX - PMIN) ) /
+    float pressure = ((((float)counts - (float)OUT_MIN_10) * (PMAX - PMIN) ) /
                         (OUT_MAX_90 - OUT_MIN_10) )
                         + PMIN;
     return pressure;
 }
 
 float abp2s_calculate_temperature(uint32_t counts) {
-    float temp = ((counts * (ABP2S_TMAX - ABP2S_TMIN)) /
+    float temp = (((float)counts * (ABP2S_TMAX - ABP2S_TMIN)) /
              (UINT24_MAX - 1))
             + ABP2S_TMIN;
     return  temp;
